Move the form to form.js
